Move the selector engine to src/internal/selectors.ts
It parses and compiles selectors and matches them against the tree. It
reads the tree through four link functions dom.ts exports and the state
predicates that already existed as functions, so no symbol crosses the
module boundary. dom.ts imports the five query functions back.
